requirements.txt
telethon==1.34.0
selenium==4.21.0
webdriver-manager==4.0.2
python-dotenv==1.1.1
